add error message for unsupported files, reset input value after upload

+++ src/js/services/conversation-service.js
@@ -6,7 +6,7 @@ import { core } from 'services/core';
 import { immediateUpdate } from 'services/user-service';
 import { initFaye } from 'utils/faye';
 import { observable } from 'utils/events';
-import { resizeImage, getBlobFromDataUrl } from 'utils/media';
+import { resizeImage, getBlobFromDataUrl, isFileTypeSupported } from 'utils/media';
 
 export function handleFirstUserMessage(response) {
     const state = store.getState();
@@ -63,38 +63,44 @@ export function sendMessage(text) {
 }
 
 export function uploadImage(file) {
-    resizeImage(file).then((dataUrl) => {
-        return sendChain(() => {
-            // add an id just to please React
-            // this message will be replaced by the real one on the server response
-            const message = {
-                mediaUrl: dataUrl,
-                mediaType: 'image/jpeg',
-                _id: Math.random(),
-                role: 'appUser',
-                status: 'sending'
-            };
+    if (isFileTypeSupported(file.type)) {
+        return resizeImage(file).then((dataUrl) => {
+            return sendChain(() => {
+                // add an id just to please React
+                // this message will be replaced by the real one on the server response
+                const message = {
+                    mediaUrl: dataUrl,
+                    mediaType: 'image/jpeg',
+                    _id: Math.random(),
+                    role: 'appUser',
+                    status: 'sending'
+                };
 
-            store.dispatch(addMessage(message));
+                store.dispatch(addMessage(message));
 
-            const user = store.getState().user;
-            const blob = getBlobFromDataUrl(dataUrl);
+                const user = store.getState().user;
+                const blob = getBlobFromDataUrl(dataUrl);
 
-            return core().conversations.uploadImage(user._id, blob, {
-                role: 'appUser'
-            }).then((response) => {
-                store.dispatch(setConversation(response.conversation));
-                observable.trigger('message:sent', response.message);
-                return response;
-            }).catch(() => {
-                store.dispatch(showErrorNotification(store.getState().ui.text.messageError));
-                store.dispatch(removeMessage({
-                    id: message._id
-                }));
+                return core().conversations.uploadImage(user._id, blob, {
+                    role: 'appUser'
+                }).then((response) => {
+                    store.dispatch(setConversation(response.conversation));
+                    observable.trigger('message:sent', response.message);
+                    return response;
+                }).catch(() => {
+                    store.dispatch(showErrorNotification(store.getState().ui.text.messageError));
+                    store.dispatch(removeMessage({
+                        id: message._id
+                    }));
 
+                });
             });
         });
-    });
+    }
+
+    store.dispatch(showErrorNotification(store.getState().ui.text.invalidFileError));
+    return Promise.reject('Invalid file type');
+
 }
 
 export function getConversation() {
